Split useSweetyMemo
