+++ src/index.ts
@@ -3,7 +3,8 @@ export type { Compare, SetSweetyState } from "./utils"
 export type { ExtractSweetyState, DeepExtractSweetyState } from "./Sweety"
 export { Sweety } from "./Sweety"
 export { batch } from "./SetStateContext"
-export { watch, useSweetyMemo } from "./watch"
+export { watch } from "./watch"
+export { useSweetyMemo } from "./useSweetyMemo"
 export { useSweety } from "./useSweety"
 export { useGetSweetyState } from "./useGetSweetyState"
 export { useSetSweetyState } from "./useSetSweetyState"
